# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking AI - API Documentation",
        "description": "API REST para el sistema de gestión de parqueaderos multi-sucursal. Permite administrar sucursales, usuarios, vehículos, tarifas, descuentos, impuestos y operaciones de entrada/salida de vehículos.",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@parking-ai.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8001/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/api/admin/document-types": {
            "get": {
                "tags": [
                    "Tipos de Documento"
                ],
                "summary": "Listar tipos de documento",
                "description": "Retorna todos los tipos de documento registrados en el sistema",
                "operationId": "73df1d69c5ab4f9ddb214b743f8f6ba3",
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Listado de tipos de documento"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID único",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Nombre del tipo de documento",
                                                        "type": "string",
                                                        "example": "Cédula de Ciudadanía"
                                                    },
                                                    "code": {
                                                        "description": "Código único",
                                                        "type": "string",
                                                        "example": "CC"
                                                    },
                                                    "description": {
                                                        "description": "Descripción opcional",
                                                        "type": "string",
                                                        "example": "Documento de identidad",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipos de Documento"
                ],
                "summary": "Crear tipo de documento",
                "description": "Registra un nuevo tipo de documento en el sistema",
                "operationId": "b1f3535151c3bd5e8892897db59a34b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre descriptivo del tipo de documento",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Tarjeta de Identidad"
                                    },
                                    "code": {
                                        "description": "Código único de identificación (debe ser único en el sistema)",
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "TI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de documento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de documento creado exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tarjeta de Identidad"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "TI"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Errores de validación"
                                        },
                                        "errors": {
                                            "properties": {
                                                "code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The code has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Laravel Sanctum. Obtén el token usando el endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}